refactor: move database commands into single file
