+++ eligibility_server/db/__init__.py
@@ -11,8 +11,7 @@ db = SQLAlchemy()
 def init_app(app):
     db.init_app(app)
 
-    from .setup import init_db_command
-    from .teardown import drop_db_command
+    from .setup import init_db_command, drop_db_command
 
     app.cli.add_command(init_db_command)
     app.cli.add_command(drop_db_command)
