+++ eligibility_server/db/setup.py
@@ -74,3 +74,22 @@ def save_users(sub: str, name: str, types: str):
     item = User(sub=sub, name=name, types=types)
     db.session.add(item)
     db.session.commit()
+
+
+@click.command("drop-db")
+def drop_db_command():
+    with current_app.app_context():
+        inspector = inspect(db.engine)
+
+        if inspector.get_table_names():
+            try:
+                click.echo(f"Users to be deleted: {User.query.count()}")
+                User.query.delete()
+                db.session.commit()
+            except Exception as e:
+                click.echo("Failed to query for Users", e)
+
+            db.drop_all()
+            click.echo("Database dropped.")
+        else:
+            click.echo("Database does not exist.")
